requirements.txt
pycairo>=1.23.0
PyGObject>=3.44.1
numpy>=1.24.0
opencv-python>=4.8.0
click>=8.1.7
